requirements.txt
websocket_client
requests
bottle